# src/api/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BadBeatMods API",
    "description": "This isn't really fully complete, but its better than absolutely nothing.\n\nThis API documentation is automatically generated and therefor may not be 100% accurate and may be missing a few fields. For example, request bodies are not fully fleshed out, and may not be accurate. Full documentation is still currently a work in progress.",
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "https://bbm.saera.gay/api"
    }
  ],
...
  "paths": {
...
    "/ba/linkVersionsExclude": {
      "post": {
        "tags": [
          "Bulk Actions"
        ],
        "summary": "",
        "description": "Submits edits if the mod is already approved, otherwise queues an edit for approval. Requires the user to be an approver.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "editIds": {
                      "type": "array",
                      "example": [
                        1,
                        2
                      ],
                      "items": {
                        "type": "number"
                      }
                    },
                    "errorIds": {
                      "type": "array",
                      "example": [
                        3
                      ],
                      "items": {
                        "type": "number"
                      }
                    },
                    "editPreformedIds": {
                      "type": "array",
                      "example": [
                        4
                      ],
                      "items": {
                        "type": "number"
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "editIds": {
                      "type": "array",
                      "example": [
                        1,
                        2
                      ],
                      "items": {
                        "type": "number"
                      }
                    },
                    "errorIds": {
                      "type": "array",
                      "example": [
                        3
                      ],
                      "items": {
                        "type": "number"
                      }
                    },
                    "editPreformedIds": {
                      "type": "array",
                      "example": [
                        4
                      ],
                      "items": {
                        "type": "number"
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "gameVersionIdFrom": {
                    "type": "number"
                  },
                  "gameVersionIdTo": {
                    "type": "number"
                  },
                  "modVersionIdsToExclude": {
                    "type": "array",
                    "items": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/ba/approveEdits": {
      "post": {
        "tags": [
          "Bulk Actions"
        ],
        "summary": "Approve multiple edit requests",
        "description": "Approve multiple edit requests. Requires the user to be an approver.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "successIds": {
                      "type": "array",
                      "example": [
                        1,
                        2
                      ],
                      "items": {
                        "type": "number"
                      }
                    },
                    "errorIds": {
                      "type": "array",
                      "example": [
                        3
                      ],
                      "items": {
                        "type": "number"
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "successIds": {
                      "type": "array",
                      "example": [
                        1,
                        2
                      ],
                      "items": {
                        "type": "number"
                      }
                    },
                    "errorIds": {
                      "type": "array",
                      "example": [
                        3
                      ],
                      "items": {
                        "type": "number"
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "bearerAuth": [],
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "approve": {
                    "type": "boolean",
                    "default": true
                  },
                  "editIds": {
                    "type": "array",
                    "items": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "bbm_session"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  }
}